Stricter task checkout requirement
This now requires that the task that gets updated at checkout:

1. has status `AVAILABLE`
2. has the correct number of previous tries

Since these two things should be updated when the task is claimed as
IN_PROGRESS, this should prevent us from claiming the same task twice.

Now, if the same task is requested twice, this will raise a
NoStatusChange. It is up to the worker to catch that and retry.